fix(obs-store): satisfy bundle export audit
